Test `health`.
The use of `setImmediate` in the unit test allows for both the `wait`
callback and the turnstile function callback to complete before
proceding to the next test so that we start each test with no occupied
turnstiles. This was added when I wanted to test the turnstile `health`
end point at the every end. I wanted to see the expected zero occupied
so I had to wait a tick. I then went back and made it uniform.