Sketch of universal arrayed response.
Create a univeral arrayed response that can be used with both `throw`
and `return`. One that preserves the notion of returning just an error
code when all you want to do is trigger an error whose status code alone
describes the error (i.e. 404).

`throw` can still be used to throw just a code and stop the request with
an error. You can still `throw` a string to affect a redirection.

Previously, the status code would have been the second argument after
the returned body. This meant you had to provide a body before the
status code. This means you can't fall on the default of using the
description as the body when all you really want to do is communicate
the error code.

Now you can return just a status code. Return just the status code and
you'll get the same behavior as if you threw the status code. If you
return just a body, that is, return a single thing that is not a number,
that is the body and the status code is OK.

If you want to return an integer as your JSON body you're going to have
to provide a status code even if the status code is the default 200.